clean up\; switching to City of shared module (almost the same information)\;adjusting City (forcing initialization)\; creating GeotoolsBootstrapping in shared module\;making ServiceData an Entity in order to allow relationship\; switching City to use refernce area\; using hibernate spatial\; forcing Id on sensor\;replacing some mock implementations with EasyMock\; fixing forcing of initialization and encapsulation in weather module; moving graph disassember, implementations and tests to util module\; using hibernate spatial for out of the box persistence of geometry data -> switching all database usage to postgres and all persistence issues to hibernate\;

+++ simulation/de.pgalise.simulation.sensorFramework/src/main/java/de/pgalise/simulation/sensorFramework/SensorRegistry.java
@@ -50,21 +50,21 @@ public interface SensorRegistry extends Iterable<Sensor>, SimulationComponent {
 	 * @param sensorId
 	 * @return sensor
 	 */
-	public Sensor getSensor(int sensorId);
+	public Sensor getSensor(long sensorId);
 
 	/**
 	 * Returns the next available id for a Sensor
 	 * 
 	 * @return the next available id for a Sensor
 	 */
-	public int nextAvailableSensorId();
+	public long nextAvailableSensorId();
 
 	/**
 	 * Returns the number of sensors in the SensorDomain
 	 * 
 	 * @return Number of sensors
 	 */
-	public int numberOfSensors();
+	public long numberOfSensors();
 
 	/**
 	 * Removes all Sensors from the SensorDomain
@@ -77,7 +77,7 @@ public interface SensorRegistry extends Iterable<Sensor>, SimulationComponent {
 	 * @param sensorId
 	 * @return Sensor which are removed
 	 */
-	public Sensor removeSensor(int sensorId);
+	public Sensor removeSensor(long sensorId);
 
 	/**
 	 * Removes a sensor from the domain
